Finishing Integration

+++ app/Services/TransactionService.php
@@ -10,19 +10,26 @@ use App\Services\Acc\Acc;
 use App\Services\Acc\AccTransaction;
 use App\Services\Acc\AccTransactionType;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Log;
 
 class TransactionService
 {
 
     protected  $acc;
     protected  $accTransaction;
+
+    protected $active;
     public function __construct(Acc $acc, AccTransaction $accTransaction) {
         $this->acc = $acc;
         $this->accTransaction = $accTransaction;
+        $this->active = env('ACC_ACTIVE');
     }
 
     public function recordSalaryToACC(SalaryRecap $data): void
     {
+        if(!$this->active){
+            return ;
+        }
         $code = "GAJIAN";
         $user = User::find($data->user_id);
         $acc  = \App\Models\Acc::where("code",$code)->first();
@@ -30,7 +37,7 @@ class TransactionService
         $transaction->type = AccTransactionType::WITHDRAWAL;
         $transaction->amount = $data->received;
         $transaction->date = $data->updated_at;
-        $transaction->description = "$code $user->name";
+        $transaction->description = "$code - $user->name";
         $transaction->source_id = $acc->source_id;
         $transaction->destination_id = $acc->destination_id;
         $transaction->tags = $code;
@@ -46,18 +53,57 @@ class TransactionService
         $data->saveQuietly();
     }
 
+    public function updateRecordSalaryToACC(SalaryRecap $data): void
+    {
+        if(!$this->active){
+            return ;
+        }
+        if($data->acc_id == null){
+            $this->recordSalaryToACC($data);
+        }
+        else{
+            $code = "GAJIAN";
+            $user = User::find($data->user_id);
+            $acc  = \App\Models\Acc::where("code",$code)->first();
+            $transaction = $this->accTransaction;
+            $transaction->type = AccTransactionType::WITHDRAWAL;
+            $transaction->amount = $data->received;
+            $transaction->date = $data->updated_at;
+            $transaction->description = "$code - $user->name";
+            $transaction->source_id = $acc->source_id;
+            $transaction->destination_id = $acc->destination_id;
+            $transaction->tags = $code;
+            $transaction->notes = $data->method;
+            $transaction->internal_reference = "ABSEN-$code-".$data->id;
+            $transaction->external_id =$data->id;
+            $this->acc->updateTransaction($data->acc_id,$transaction);
+        }
+    }
+
+    public function deleteRecordSalaryToACC(SalaryRecap $data)
+    {
+        if(!$this->active){
+            return ;
+        }
+        if($data->acc_id){
+            $this->acc->delete($data->acc_id);
+        }
+    }
+
 
     public function recordPayLoanACC(LoanPayment $data): void
     {
-        $time = Carbon::now()->format("H:i:s");
+        if(!$this->active){
+            return ;
+        }
         $code = "BAYARKASBON";
         $user = User::find($data->user_id);
         $acc  = \App\Models\Acc::where("code",$code)->first();
         $transaction = $this->accTransaction;
         $transaction->type = AccTransactionType::DEPOSIT;
         $transaction->amount = $data->amount;
-        $transaction->date = $data->date." ".$time;
-        $transaction->description = "$code $user->name";
+        $transaction->date = $data->date;
+        $transaction->description = "$code - $user->name";
         $transaction->source_id = $acc->source_id;
         $transaction->destination_id = $acc->destination_id;
         $transaction->tags = $code;
@@ -75,19 +121,24 @@ class TransactionService
 
     public function updateRecordPayLoanACC(LoanPayment $data): void
     {
+        if(!$this->active){
+            return ;
+        }
         if($data->acc_id == null){
             $this->recordPayLoanACC($data);
         }
 
         else{
             $code = "BAYARKASBON";
-            $time = Carbon::now()->format("H:i:s");
             $user = User::find($data->user_id);
+            $acc  = \App\Models\Acc::where("code",$code)->first();
 
             $transaction = $this->accTransaction;
             $transaction->amount = $data->amount;
-            $transaction->description = "$code $user->name";
-            $transaction->date = $data->date." ".$time;
+            $transaction->description = "$code - $user->name";
+            $transaction->date = $data->date;
+            $transaction->source_id = $acc->source_id;
+            $transaction->destination_id = $acc->destination_id;
 
             $this->acc->updateTransaction($data->acc_id,$transaction);
         }
@@ -96,6 +147,9 @@ class TransactionService
 
     public function deleteRecordPayLoanAcc(LoanPayment $data)
     {
+        if(!$this->active){
+            return ;
+        }
         if($data->acc_id){
             $this->acc->delete($data->acc_id);
         }
@@ -110,16 +164,18 @@ class TransactionService
 
     public function recordLoanACC(Loan $loan)
     {
+        if(!$this->active){
+            return ;
+        }
         $code = "KASBON";
-        $time = Carbon::now()->format("H:i:s");
         $user = User::find($loan->user_id);
         $acc  = \App\Models\Acc::where("code",$code)->first();
 
         $transaction = $this->accTransaction;
         $transaction->type = AccTransactionType::DEPOSIT;
         $transaction->amount = $loan->amount;
-        $transaction->date = $loan->date." ".$time;
-        $transaction->description = "$code $user->name";
+        $transaction->date = $loan->date;
+        $transaction->description = "$code - $user->name";
         $transaction->source_id = $acc->source_id;
         $transaction->destination_id = $acc->destination_id;
         $transaction->tags = $code;
@@ -135,9 +191,12 @@ class TransactionService
 
     public function updateRecordLoanACC(Loan $loan)
     {
+        if(!$this->active){
+            return ;
+        }
         $code = "KASBON";
-        $time = Carbon::now()->format("H:i:s");
         $user = User::find($loan->user_id);
+        $acc  = \App\Models\Acc::where("code",$code)->first();
 
         if($loan->acc_id == null){
             $this->recordLoanACC($loan);
@@ -145,14 +204,20 @@ class TransactionService
         else{
             $transaction = $this->accTransaction;
             $transaction->amount = $loan->amount;
-            $transaction->description = "$code $user->name";
-            $transaction->date = $loan->date." ".$time;
+            $transaction->description = "$code - $user->name";
+            $transaction->date = $loan->date;
+            $transaction->source_id = $acc->source_id;
+            $transaction->destination_id = $acc->destination_id;
             $this->acc->updateTransaction($loan->acc_id, $transaction);
         }
     }
 
     public function deleteRecordLoanACC(Loan $loan)
     {
+        if(!$this->active){
+            return ;
+        }
+
         if($loan->acc_id){
             $this->acc->delete($loan->acc_id);
         }
